requirements.txt
ttkbootstrap>=1.20.2
Pillow>=12.1.1
matplotlib>=3.10.8
pywindnd>=1.0.9